Create higgs_webui_complete.py
Higgs Audio V2 WebUI
Mostly "vibe-coded" trash, but works for my purposes. Maybe it'll work for you, too. I took a bunch of scripts I wrote for Higgs, the original generation script, and attempted to marry them in a semi-functional way.

For use with the quantized model project. Put in examples/ and run to launch the WebUI.

Put your voice samples in voice_prompts/